fix float

+++ randresist/randresist.py
@@ -5,7 +5,7 @@ import random
 rvals = np.array([float(x) for x in "1.0 1.1 1.2 1.3 1.5 1.6 1.8 2.0 2.2 2.4 2.7 3.0 3.3 3.6 3.9 4.3 4.7 5.1 5.6 6.2 6.8 7.5 8.2 9.1".split()])
 
 def getRList(low, high):
-    vals = np.empty(0,np.float)
+    vals = np.empty(0,float)
 
     l10 = np.log(10)
     logLow = int(np.floor(np.log(low)/l10 + 0.0001))
